Moving math out of basegl
