+++ src/rust/ensogl/src/lib.rs
@@ -56,10 +56,10 @@ pub mod data;
 pub mod debug;
 pub mod display;
 pub mod gui;
-pub mod math;
 pub mod system;
 
-pub use enso_frp as frp;
+pub use enso_frp  as frp;
+pub use enso_math as math;
 
 /// Prelude - commonly used utilities.
 pub mod prelude {
