+++ src/rust/lib/math/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name    = "enso-math"
+version = "0.1.0"
+authors = ["Enso Team <contact@luna-lang.org>"]
+edition = "2018"
+
+[lib]
+
+[dependencies]
+nalgebra   = { version = "0.19.0" }
+num-traits = { version = "0.2" }
