+++ src/rust/lib/math/src/geometry.rs
