+++ src/rust/lib/math/src/topology.rs
